Version 2.0 Final
Log Notes

Version 2.0 - iOS/Android (January 5, 2016)

We are proud to introduce AggiesLand 2.0. This update add features that are missing in first version of AggiesLand, such as uploading events with pictures. Also with this update, this address all bugs and problems with the first app. We have also revive the twitter feed that was missing so you can check realtime tweets from NC A&T and around Greensboro.

Features include:
	-FEATURE: Twitter Feed
	-FEATURE: Upload Events with existing photos.
	-FEATURE: Profile Viewer: You can upload you picture and check you status on the side 		        bar.

Bug Fixes:
	-FIXED: Not able to go back from Bus Routes or Sports page
	-FIXED: Major UI Design Changes
	-FIXED: Other stability improvements

If you have any problems with the app, please email us at support@aggiesland.me